requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.9.0
python-dotenv>=1.0.0
alembic>=1.8.0
